requirements.txt
PyYAML~=5.3.1
keyboard~=0.13.5
psutil~=5.7.0